Move get_user_from_saml_attributes to saml helpers
Move the function over to the saml_helpers file and start a saml_helpers test suite to prepare to put the other functions under test
